requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
gunicorn==20.1.0
mongoengine==0.23.1
pymongo==3.12.0
python-snappy==0.6.0
pytz==2021.1
sqlparse==0.2.4
whitenoise==5.3.0
